Refactor: abstracted function, made url source of truth for everything
