+++ src/pages/graveyard.js
@@ -1,91 +1,13 @@
 import {
-  projectCard,
-  projectCardSkeleton,
-} from "../components/features/projectCard/projectCard.js";
-import { pagination } from "../components/features/pagination/pagination.js";
-import { api } from "../main.js";
+  setPageState,
+  setupGraveyardForms,
+  fetchAndRenderProjects,
+} from "../components/features/graveyard/index.js";
 
-function getPageState() {
-  const params = new URLSearchParams(window.location.search);
-
-  let limit = Number(params.get("limit")) || 10;
-  limit = Math.min(100, Math.max(1, limit));
-
-  const offset = Number(params.get("offset")) || 0;
-  const currentPage = Math.floor(offset / limit) + 1;
-
-  return {
-    order: params.get("order") || "desc",
-    orderBy: params.get("orderBy") || "createdAt",
-    query: params.get("query") || "",
-    types: params.get("types") || "",
-    limit,
-    offset,
-    currentPage,
-  };
-}
-let { order, orderBy, query, types, limit, offset, currentPage } =
-  getPageState();
-
-const projectsContainer = document.querySelector(".project-container");
-const paginationContainer = document.querySelector(".pagination");
-
-function showSkeletons(count = limit) {
-  projectsContainer.innerHTML = "";
-  for (let i = 0; i < count; i++) {
-    projectsContainer.innerHTML += projectCardSkeleton();
-  }
-}
-
-async function fetchAndRenderProjects(page = 1) {
-  showSkeletons();
-  try {
-    offset = (page - 1) * limit;
-    const url = new URL(window.location);
-    url.searchParams.set("offset", offset);
-    window.history.pushState({}, "", url);
-    let options = { offset, limit, types, order, orderBy };
-    query?.length ? (options.query = query) : delete options.query;
-    const res = await api.getAllProjects(options);
-
-    console.log(res);
-
-    const projects = res.data || [];
-    const total = res.meta?.total || 0;
-    const totalPages = Math.ceil(total / limit);
-
-    projectsContainer.innerHTML = "";
-
-    if (projects.length === 0) {
-      projectsContainer.innerHTML =
-        "<p>No projects found. Try another page.</p>";
-    } else {
-      projects.forEach((project) => {
-        projectsContainer.append(projectCard(project));
-      });
-    }
-
-    pagination({
-      current: currentPage,
-      total: totalPages,
-      container: paginationContainer,
-      onPageChange: (newPage) => {
-        currentPage = newPage;
-        fetchAndRenderProjects(newPage);
-      },
-    });
-  } catch (err) {
-    console.log(err);
-    projectsContainer.innerHTML =
-      "<p>An error occurred fetching projects, please refresh the page.</p>";
-  }
-}
-
-fetchAndRenderProjects(currentPage);
+fetchAndRenderProjects();
 
 /* Modal for filter */
 const dialog = document.querySelector("dialog");
-const typesCheckboxes = document.querySelector("#types");
 const dialogCloseBtn = document.querySelector("#close-dialog");
 const dialogOpenBtn = document.querySelector("#open-dialog");
 dialogOpenBtn.addEventListener("click", () => {
@@ -97,74 +19,24 @@ dialogCloseBtn.addEventListener("click", () => {
 });
 
 /* Form Setup */
-const orderSelector = document.querySelector("#order-selector");
-orderSelector.value = order;
-const orderBySelector = document.querySelector("#orderby-selector");
-orderBySelector.value = orderBy;
-const limitSelector = document.querySelector("#limit-selector");
-limitSelector.value = limit;
 
-const typesRes = await api.getAllTypes();
-let typesArray = [];
-if (typesRes.success) {
-  typesArray = typesRes.data;
-}
-
-typesArray.forEach((type) => {
-  const label = document.createElement("label");
-  label.setAttribute("for", `type-${type.id}`);
-  label.style.display = "block"; // if you want each on a new line
-
-  const checkbox = document.createElement("input");
-  checkbox.type = "checkbox";
-  checkbox.name = "types";
-  checkbox.value = type.id;
-  checkbox.id = `type-${type.id}`;
-
-  const selectTypes = types.split(",");
-  if (selectTypes.includes(`${type.id}`)) {
-    checkbox.checked = true;
-  }
-
-  label.appendChild(checkbox);
-  label.append(` ${type.name}`);
-  typesCheckboxes.appendChild(label);
-});
+setupGraveyardForms();
 
 /* Form Handler */
 const filterForm = document.getElementById("project-filters");
 const searchForm = document.getElementById("search-form");
-searchForm.query.value = query;
 async function formHandler(e) {
   e.preventDefault();
   const formData = new FormData(filterForm);
-  let searchQuery = searchForm.query.value;
   const data = Object.fromEntries(formData.entries());
+  data.types = formData.getAll("types").join(",");
+  const searchQuery = searchForm.query.value.trim();
+  if (searchQuery) data.query = searchQuery;
 
-  const selectedTypes = formData.getAll("types").join(",");
-  data.types = selectedTypes;
-  if (searchQuery.trim().length > 0) {
-    data.query = searchQuery;
-  } else {
-    delete data.query;
-  }
-  console.log(data.query);
-  const url = new URL(window.location);
-  url.searchParams.forEach((_, key) => url.searchParams.delete(key));
-  for (const [key, value] of Object.entries(data)) {
-    url.searchParams.set(key, value);
-  }
-
-  window.history.pushState({}, "", url);
-  order = data.order;
-  orderBy = data.orderBy;
-  query = data.query;
-  types = selectedTypes;
-  limit = data.limit;
-  currentPage = 1;
-
-  fetchAndRenderProjects(currentPage);
-  dialog.close();
+  setPageState(data, () => {
+    fetchAndRenderProjects();
+    dialog.close();
+  });
 }
 
 filterForm.addEventListener("submit", async (e) => {
@@ -184,13 +56,9 @@ clearFilterBtn.addEventListener("click", async (e) => {
 const clearSearchBtn = document.querySelector("#clear-search");
 clearSearchBtn.addEventListener("click", async (e) => {
   searchForm.reset();
-  query = "";
   formHandler(e);
-  console.log(query);
 });
 
 addEventListener("popstate", () => {
-  [order, orderBy, query, types, limit, offset, currentPage] =
-    Object.values(getPageState());
-  fetchAndRenderProjects(currentPage);
+  fetchAndRenderProjects();
 });
